backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.0
PyJWT==2.8.0
mediapipe==0.10.7
Pillow==10.1.0
